+++ internal/rest/server_proposals.go
@@ -208,7 +208,7 @@ func (s *Server) listProposals(w http.ResponseWriter, r *http.Request) {
 		Limit:    limit,
 		Dao:      f.DAO,
 		Category: string(f.Category),
-		Title:    f.Title,
+		Title:    f.Query,
 	})
 	if err != nil {
 		log.Error().Err(err).Msg("get proposal list")